refactored doc layout
