+++ docs/src/SUMMARY.md
@@ -7,10 +7,8 @@
 # Instruments
 - [Instruments](./instruments/instruments.md)
 - [Test Daq](./instruments/test_daq.md)
-- [Test Cryostat](./instruments/test_cryostat.md)
 - [Keithley2400](./instruments/keithley2400.md)
 - [Siglentsds2352Xe](./instruments/siglentsds2352xe.md)
-- [Test Spectrometer](./instruments/test_spectrometer.md)
 - [Scryostation](./instruments/cryostats/scryostation.md)
 # Additional Information
 - [Information](./information/information.md)
