--- docs/src/instruments/test_cryostat.md
@@ -1,32 +1,0 @@
-# Test_cryostat
-
-
-## Methods
-
-### setup_config
-
-
-
-### measure
-
-
-
-### goto_setpoint
-
-**Signature:** `goto_setpoint(setpoint)`
-
-
-
-### set_magneticfield
-
-**Signature:** `set_magneticfield(strength)`
-
-
-
-### get_cryostate
-
-
-
-### get_magnetstate
-
-
