--- docs/src/instruments/test_spectrometer.md
@@ -1,18 +1,0 @@
-# Test_spectrometer
-
-
-## Methods
-
-### setup_config
-
-
-
-### measure
-
-
-
-### goto_wavelength
-
-**Signature:** `goto_wavelength(wavelength)`
-
-
