+++ src/spcs_instruments/instruments/__init__.py
@@ -1,7 +1,7 @@
 from .test_daq import Test_daq
 from .siglent_sds_2352xe_driver import SiglentSDS2352XE
 from .keithley_2400_driver import Keithley2400
-from .test_cryostat import Test_cryostat
+from .cryostats import Test_cryostat
 from .cryostats import Scryostation
 from .spectrometers import HoribaiHR550,Test_spectrometer 
 __all__ = ["Test_cryostat", "Test_daq", "SiglentSDS2352XE", "Keithley2400", "Test_spectrometer", "Scryostation", "HoribaiHR550"]
