+++ src/spcs_instruments/instruments/cryostats/__init__.py
@@ -1,3 +1,3 @@
 from .scryostation_driver import Scryostation
-
-__all__ = ["Scryosation"]
+from .test_cryostat import Test_cryostat
+__all__ = ["Scryosation", "Test_cryostat"]
